requirements.txt
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
Markdown==3.3.7
gunicorn==20.1.0
